+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   [ksuther](https://github.com/ksuther)
   [#291](https://github.com/SlatherOrg/slather/pull/291)
 
+* Add `--configuration` option
+  [thasegaw](https://github.com/thasegaw)
+  [#xxx](https://github.com/slatherOrg/slather/pull/xxx)
+
 * Automatically ignore headers in Xcode platform SDKs.  
   [ksuther](https://github.com/ksuther)
   [#286](https://github.com/SlatherOrg/slather/pull/286)